+++ Core/Executor/ContentTypeGroupManager.php
@@ -134,14 +134,16 @@ class ContentTypeGroupManager extends RepositoryExecutor implements MigrationGen
     /**
      * @param ContentTypeGroup $object
      * @param array $references the definitions of the references to set
-     * @throws \InvalidArgumentException When trying to assign a reference to an unsupported attribute
+     * @throws InvalidStepDefinitionException
      * @return array key: the reference names, values: the reference values
      */
     protected function getReferencesValues($object, array $references, $step)
     {
         $refs = array();
 
-        foreach ($references as $reference) {
+        foreach ($references as $key => $reference) {
+
+            $reference = $this->parseReferenceDefinition($key, $reference);
 
             switch ($reference['attribute']) {
                 case 'content_type_group_id':
@@ -153,7 +155,7 @@ class ContentTypeGroupManager extends RepositoryExecutor implements MigrationGen
                     $value = $object->identifier;
                     break;
                 default:
-                    throw new \InvalidArgumentException('Content Type Group Manager does not support setting references for attribute ' . $reference['attribute']);
+                    throw new InvalidStepDefinitionException('Content Type Group Manager does not support setting references for attribute ' . $reference['attribute']);
             }
 
             $refs[$reference['identifier']] = $value;