+++ Core/Executor/LocationManager.php
@@ -295,14 +295,17 @@ class LocationManager extends RepositoryExecutor
     /**
      * @param Location $location
      * @param array $references the definitions of the references to set
-     * @throws \InvalidArgumentException When trying to assign a reference to an unsupported attribute
+     * @throws InvalidStepDefinitionException
      * @return array key: the reference names, values: the reference values
      */
     protected function getReferencesValues($location, array $references, $step)
     {
         $refs = array();
 
-        foreach ($references as $reference) {
+        foreach ($references as $key => $reference) {
+
+            $reference = $this->parseReferenceDefinition($key, $reference);
+
             switch ($reference['attribute']) {
                 case 'location_id':
                 case 'id':
@@ -379,7 +382,7 @@ class LocationManager extends RepositoryExecutor
                     $value = $this->sortConverter->sortOrder2Hash($location->sortOrder);
                     break;
                 default:
-                    throw new \InvalidArgumentException('Location Manager does not support setting references for attribute ' . $reference['attribute']);
+                    throw new InvalidStepDefinitionException('Location Manager does not support setting references for attribute ' . $reference['attribute']);
             }
 
             $refs[$reference['identifier']] = $value;