+++ Core/Executor/SectionManager.php
@@ -137,14 +137,15 @@ class SectionManager extends RepositoryExecutor implements MigrationGeneratorInt
     /**
      * @param Section $section
      * @param array $references the definitions of the references to set
-     * @throws \InvalidArgumentException When trying to assign a reference to an unsupported attribute
+     * @throws InvalidStepDefinitionException
      * @return array key: the reference names, values: the reference values
      */
     protected function getReferencesValues($section, array $references, $step)
     {
         $refs = array();
 
-        foreach ($references as $reference) {
+        foreach ($references as $key => $reference) {
+            $reference = $this->parseReferenceDefinition($key, $reference);
             switch ($reference['attribute']) {
                 case 'section_id':
                 case 'id':
@@ -159,7 +160,7 @@ class SectionManager extends RepositoryExecutor implements MigrationGeneratorInt
                     $value = $section->name;
                     break;
                 default:
-                    throw new \InvalidArgumentException('Section Manager does not support setting references for attribute ' . $reference['attribute']);
+                    throw new InvalidStepDefinitionException('Section Manager does not support setting references for attribute ' . $reference['attribute']);
             }
 
             $refs[$reference['identifier']] = $value;